refactor(popup): move time saved into a new component
